+++ .github/workflows/root-ci-config/build_utils.py
@@ -8,6 +8,7 @@ import textwrap
 from functools import wraps
 from http import HTTPStatus
 from typing import Callable, Dict
+from collections import namedtuple
 
 from openstack.connection import Connection
 from requests import get
@@ -108,6 +109,28 @@ def subprocess_with_log(command: str) -> int:
     log.add(command)
 
     return result.returncode
+
+def subprocess_with_capture(command: str):
+    """Runs <command> in shell, capture output and appends <command> to log"""
+
+    print_fancy(textwrap.dedent(command), sgr=1)
+
+    print("\033[90m", end='')
+
+    if os.name == 'nt':
+        command = "$env:comspec = 'cmd.exe'; " + command
+
+    result = subprocess.run(command, capture_output=True, text=True, shell=True, check=False)
+
+    print(result.stdout)
+    print(result.stderr)
+    print("\033[0m", end='')
+
+    # Since we are capturing the result and using it in other command later,
+    # we don't need it for the reproducing steps.
+    # So no call to: log.add(command)
+
+    return result
 
 
 def die(code: int = 1, msg: str = "") -> None: