+++ core/foundation/inc/ROOT/RVersion.hxx
@@ -4,8 +4,8 @@
 /* Update on release: */
 #define ROOT_VERSION_MAJOR 6
 #define ROOT_VERSION_MINOR 30
-#define ROOT_VERSION_PATCH 04
-#define ROOT_RELEASE_DATE "Jan 31 2024"
+#define ROOT_VERSION_PATCH 06
+#define ROOT_RELEASE_DATE "Apr 3 2024"
 
 /* Don't change the lines below. */