+++ roofit/xroofit/src/xRooNode.cxx
@@ -16,6 +16,30 @@
 The xRooNode class is designed to wrap over a TObject and provide functionality to aid with interacting with that
 object, particularly in the case where the object is a RooFit class instance. It is a smart pointer to the object, so
 you have access to all the methods of the object too.
+
+xRooNode is designed to work in both python and C++, but examples below are given in python because that is imagined
+  be the most common way to use the xRooFit API.
+
+-# [Exploring workspaces](\ref exploring-workspaces)
+
+\anchor exploring-workspaces
+## Exploring workspaces
+
+An existing workspace file (either a ROOT file containing a RooWorkspace, or a json HS3 file) can be opened using
+ xRooNode like this:
+
+\code{.py}
+from ROOT.Experimental import XRooFit
+w = XRooFit.xRooNode("workspace.root") # or can use workspace.json for HS3
+\endcode
+
+ You can explore the content of the workspace somewhat like you would a file system: each node contains sub-nodes,
+ which you can interact with to explore ever deeper. The most relevant methods for navigating the workspace and exploring
+ the content are:
+
+
+
+
  */
 
 #include "RVersion.h"
@@ -183,6 +207,8 @@ auto GETLISTTREE(TGFileBrowser *b)
 #include "RooNaNPacker.h"
 #endif
 
+
+
 BEGIN_XROOFIT_NAMESPACE;
 
 xRooNode::InteractiveObject *xRooNode::gIntObj = nullptr;
@@ -200,6 +226,14 @@ const T &_or_func(const T &a, const T &b)
       return a;
    return b;
 }
+
+////////////////////////////////////////////////////////////////////////////////
+/// Create new object of type classname, with given name and title, and own-wrap it
+/// i.e. the xRooNode will delete the object when the node (and any that reference it) is destroyed
+///
+/// \param classname : the type of the object to create
+/// \param name : the name to give the object
+/// \param title : the title to give the object
 
 xRooNode::xRooNode(const char *classname, const char *name, const char *title)
    : xRooNode(name, std::shared_ptr<TObject>(TClass::GetClass(classname)
@@ -1489,7 +1523,8 @@ xRooNode xRooNode::Add(const xRooNode &child, Option_t *opt)
             return xRooNode(*_ws->data(asi.first->GetName()), fParent);
          }
 
-         auto _obs = fParent->obs().argList();
+         auto parentObs = fParent->obs(); // may own globs so keep alive
+         auto _obs = parentObs.argList();
          // put globs in a snapshot
          std::unique_ptr<RooAbsCollection> _globs(_obs.selectByAttrib("global", true));
          // RooArgSet _tmp; _tmp.add(*_globs);_ws->saveSnapshot(child.GetName(),_tmp);
@@ -2008,9 +2043,18 @@ xRooNode xRooNode::Add(const xRooNode &child, Option_t *opt)
    } else if (auto w = get<RooWorkspace>(); w) {
       child.convertForAcquisition(*this);
       if (child.get()) {
-         auto out = acquire(child.fComp);
-         if (out)
-            return xRooNode(child.GetName(), out, *this);
+         if(auto _d = child.get<RooAbsData>()) {
+            // don't use acquire method to import, because that adds datasets as Embeddded
+            if (!w->import(*_d)) {
+               return xRooNode(child.GetName(), *w->data(child.GetName()),*this);
+            } else {
+               throw std::runtime_error(TString::Format("Could not import dataset %s into workspace %s",child.GetName(),w->GetName()).Data());
+            }
+         } else {
+            auto out = acquire(child.fComp);
+            if (out)
+               return xRooNode(child.GetName(), out, *this);
+         }
       }
 
       if (!child.empty() || child.fFolder == "!models") {
@@ -4045,6 +4089,9 @@ std::shared_ptr<xRooNode> xRooNode::at(const std::string &name, bool browseResul
    return res;
 }
 
+////////////////////////////////////////////////////////////////////////////////
+/// The RooWorkspace this node belong to, if any
+
 RooWorkspace *xRooNode::ws() const
 {
    if (auto _w = get<RooWorkspace>(); _w)
@@ -5058,6 +5105,9 @@ xRooNode &xRooNode::browse()
    return *this;
 }
 
+////////////////////////////////////////////////////////////////////////////////
+/// List of observables (global and regular) of this node.
+
 xRooNode xRooNode::obs() const
 {
    xRooNode out(".obs", std::make_shared<RooArgList>(), *this);
@@ -5070,6 +5120,9 @@ xRooNode xRooNode::obs() const
    }
    return out;
 }
+
+////////////////////////////////////////////////////////////////////////////////
+/// List of global observables of this node.
 
 xRooNode xRooNode::globs() const
 {
@@ -5084,6 +5137,9 @@ xRooNode xRooNode::globs() const
    return out;
 }
 
+////////////////////////////////////////////////////////////////////////////////
+/// List of regular observables of this node.
+
 xRooNode xRooNode::robs() const
 {
    xRooNode out(".robs", std::make_shared<RooArgList>(), *this);
@@ -5096,6 +5152,9 @@ xRooNode xRooNode::robs() const
    }
    return out;
 }
+
+////////////////////////////////////////////////////////////////////////////////
+/// List of parameters (non-observables) of this node.
 
 xRooNode xRooNode::pars() const
 {
@@ -5110,6 +5169,9 @@ xRooNode xRooNode::pars() const
    return out;
 }
 
+////////////////////////////////////////////////////////////////////////////////
+/// List of parameters that are currently constant
+
 xRooNode xRooNode::consts() const
 {
    xRooNode out(".consts", std::make_shared<RooArgList>(), *this);
@@ -5122,6 +5184,10 @@ xRooNode xRooNode::consts() const
    }
    return out;
 }
+
+////////////////////////////////////////////////////////////////////////////////
+/// List of parameters that are currently non-constant
+/// These parameters do not have the "Constant" attribute
 
 xRooNode xRooNode::floats() const
 {
@@ -5136,6 +5202,10 @@ xRooNode xRooNode::floats() const
    return out;
 }
 
+////////////////////////////////////////////////////////////////////////////////
+/// List of parameters of interest: parameters marked as "of interest"
+/// These parameters have the "poi" attribute
+
 xRooNode xRooNode::poi() const
 {
    xRooNode out(".poi", std::make_shared<RooArgList>(), *this);
@@ -5148,6 +5218,10 @@ xRooNode xRooNode::poi() const
    }
    return out;
 }
+
+////////////////////////////////////////////////////////////////////////////////
+/// List of nuisance parameters: non-constant parameters that are not marked of interest,
+///  as well as any parameters that have been marked by the "np" attribute
 
 xRooNode xRooNode::np() const
 {
@@ -5164,6 +5238,9 @@ xRooNode xRooNode::np() const
    return out;
 }
 
+////////////////////////////////////////////////////////////////////////////////
+/// List of prespecified parameters: non-floatable parameters
+
 xRooNode xRooNode::pp() const
 {
    xRooNode out(".pp", std::make_shared<RooArgList>(), *this);
@@ -5177,6 +5254,9 @@ xRooNode xRooNode::pp() const
    }
    return out;
 }
+
+////////////////////////////////////////////////////////////////////////////////
+/// List of variables (observables and parameters) of this node
 
 xRooNode xRooNode::vars() const
 {
@@ -5411,6 +5491,9 @@ xRooNode xRooNode::components() const
    return out;
 }
 
+////////////////////////////////////////////////////////////////////////////////
+/// bins of a channel or sample, or channels of a multi-channel pdf
+
 xRooNode xRooNode::bins() const
 {
    xRooNode out(".bins", nullptr, *this);
@@ -5596,7 +5679,8 @@ xRooNode xRooNode::factors() const
       // if workspace, return all functions (not pdfs) that have a RooProduct as one of their clients
       // or not clients
       // exclude obs and globs
-      auto _obs = obs().argList();
+      auto oo = obs(); // need to keep alive as may contain owning globs
+      auto& _obs = *(oo.get<RooArgList>());
       for (auto a : w->allFunctions()) {
          if (_obs.contains(*a))
             continue;
@@ -5736,7 +5820,8 @@ xRooNode xRooNode::datasets() const
           (!get() && fParent &&
            fParent->get<RooAbsPdf>())) { // second condition handles 'bins' nodes of pdf, which have null ptr
          // only add datasets that have observables that cover all our observables
-         RooArgSet _obs(obs().argList());
+         auto oo = obs(); // must keep alive in case is owning the globs
+         RooArgSet _obs(*oo.get<RooArgList>());
          //_obs.add(coords(true).argList(), true); // include coord observables too, and current xaxis if there's one -
          // added in loop below
 
@@ -5773,14 +5858,14 @@ xRooNode xRooNode::datasets() const
                   hasMissing = true;
                } else {
                   if (extraCut != "")
-                     extraCut += " && ";
+                     extraCut += " || ";
                   extraCut += TString::Format("%s==%d", s->indexCat().GetName(), cat.second);
                }
             }
             if (hasMissing) {
                if (cut != "")
                   cut += " && ";
-               cut += extraCut;
+               cut += "(" + extraCut + ")";
                cutobs.add(s->indexCat());
             }
          }
@@ -5799,9 +5884,9 @@ xRooNode xRooNode::datasets() const
                   // TODO: Could consider using a 'filter' node (see filter() method) applied to the dataset instead
                   // of creating and using a reduced dataset here
                   out.emplace_back(std::make_shared<xRooNode>(
-                     std::shared_ptr<RooAbsData>(d->get<RooAbsData>()->reduce(cutFormula)), *this));
+                     std::shared_ptr<RooAbsData>(d->get<RooAbsData>()->reduce(*std::unique_ptr<RooAbsCollection>(d->robs().get<RooArgList>()->selectCommon(_obs)),cutFormula)), *this));
                   // put a subset of the globs in the returned dataset too
-                  out.back()->get<RooAbsData>()->setGlobalObservables(*globs().get<RooArgList>());
+                  out.back()->get<RooAbsData>()->setGlobalObservables(*std::unique_ptr<RooAbsCollection>(d->globs().get<RooArgList>()->selectCommon(*globs().get<RooArgList>())));
                   if (d->get()->TestBit(1 << 20))
                      out.back()->get()->SetBit(1 << 20);
                   // need to attach the original dataset so that things like SetBinContent can interact with it
@@ -5872,10 +5957,18 @@ TGraph *xRooNode::BuildGraph(RooAbsLValue *v, bool includeZeros, TVirtualPad *fr
                      theHist->GetXaxis()->SetBinLabel(i + 1, cat->getLabel());
                   }
                } else {
-                  theHist = new TH1D(
-                     TString::Format("%s_%s", GetName(), vo->GetName()),
-                     TString::Format("my temp hist;%s", strlen(vo->GetTitle()) ? vo->GetTitle() : vo->GetName()),
-                     v->numBins(), v->getBinningPtr(nullptr)->lowBound(), v->getBinningPtr(nullptr)->highBound());
+                  auto _binning = v->getBinningPtr(nullptr);
+                  if(_binning->isUniform()) {
+                     theHist = new TH1D(
+                        TString::Format("%s_%s", GetName(), vo->GetName()),
+                        TString::Format("my temp hist;%s", strlen(vo->GetTitle()) ? vo->GetTitle() : vo->GetName()),
+                        v->numBins(), _binning->lowBound(), _binning->highBound());
+                  } else {
+                     theHist = new TH1D(
+                        TString::Format("%s_%s", GetName(), vo->GetName()),
+                        TString::Format("my temp hist;%s", strlen(vo->GetTitle()) ? vo->GetTitle() : vo->GetName()),
+                        v->numBins(), _binning->array());
+                  }
                }
             } else {
                throw std::runtime_error("Cannot draw dataset without parent PDF");
@@ -5983,7 +6076,8 @@ TGraph *xRooNode::BuildGraph(RooAbsLValue *v, bool includeZeros, TVirtualPad *fr
                                 (xvar && val) ? xPos->GetBinContent(i + 1) : theHist->GetBinCenter(i + 1), val);
 
             // x-error will be the (weighted) standard deviation of the x values ...
-            double xErr = sqrt(xPos2->GetBinContent(i + 1) - pow(xPos->GetBinContent(i + 1), 2));
+            double xErr = xPos2->GetBinContent(i + 1) - pow(xPos->GetBinContent(i + 1), 2);
+            xErr = (xErr <= 0) ? 0. : sqrt(xErr); // protects against floating point rounding effects
 
             if (xErr || val)
                dataGraph->SetPointError(dataGraph->GetN() - 1, xErr, xErr,
@@ -8079,6 +8173,9 @@ public:
          getLegend(false, true);
          fPad->GetCanvas()->Paint();
          fPad->GetCanvas()->Update();
+#if ROOT_VERSION_CODE >= ROOT_VERSION(6, 30, 00)
+         fPad->GetCanvas()->ResetUpdated(); // stops previous canvas being replaced in a jupyter notebook
+#endif
          fPad->cd();
       }
       nExisting--;
@@ -8122,9 +8219,8 @@ void xRooNode::Draw(Option_t *opt)
       if (gPad)
          gPad->Clear();
       xRooNLLVar::xRooHypoPoint(std::dynamic_pointer_cast<RooStats::HypoTestResult>(fComp)).Draw(opt);
-      if (gPad) {
-         gPad->GetCanvas()->Paint();
-         gPad->GetCanvas()->Update();
+      {
+         PadRefresher p(gPad); // refreshes the pad
       }
       gSystem->ProcessEvents();
       return;
@@ -10510,6 +10606,34 @@ std::vector<double> xRooNode::GetBinErrors(int binStart, int binEnd, const xRooN
          res *= ax->GetBinWidth(bin);
       }
       out.push_back(res);
+   }
+
+   return out;
+}
+
+
+   std::string cling::printValue( const xRooNode *v ) {
+   if(!v) return "nullptr\n";
+   if(!v->empty()) {
+      std::string out;
+      size_t left = v->size();
+      for(auto n : *v) {
+         left--;
+         if(!out.empty()) out += ","; else out += "{";
+         out += n->GetName();
+         if(out.length()>100 && left > 0) {
+            out += TString::Format(",... and %lu more",left);
+            break;
+         }
+      }
+      out += "}\n";
+      return out;
+   }
+   std::string out;
+   if(!(*v)) {
+      return "<empty node>";
+   } else {
+      return Form("Name: %s",v->GetName());
    }
 
    return out;